move all config code to preferences
